Some refactoring of the kmer indexing
