+++ storage/variant/io/__init__.py
@@ -16,15 +16,6 @@ def check_variants_table_columns(df: pd.DataFrame) -> None:
                         f'Expected {expected_columns}, actual {actual_columns}')
 
 
-def execute_commands(commands: List[List[str]]):
-    try:
-        for command in commands:
-            subprocess.run(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True, text=True)
-    except subprocess.CalledProcessError as e:
-        err_msg = str(e.stderr.strip())
-        raise Exception(f'Could not run [{" ".join(e.cmd)}]: error {err_msg}')
-
-
 class VariantsReader(abc.ABC):
 
     def __init(self):
